migrating cc function to separate file
